Fix circulars and move cancel functions into one file
