+++ src/mahoji/commands/ge.ts
@@ -5,8 +5,7 @@ import { sumArr, uniqueArr } from 'e';
 import { CommandRunOptions } from 'mahoji';
 import { CommandOption } from 'mahoji/dist/lib/types';
 
-import { createGECancelButton } from '../../lib/geHelpers';
-import { GrandExchange } from '../../lib/grandExchange';
+import { createGECancelButton, GrandExchange } from '../../lib/grandExchange';
 import { marketPricemap } from '../../lib/marketPrices';
 import { prisma } from '../../lib/settings/prisma';
 import { formatDuration, itemNameFromID, makeComponents, toKMB } from '../../lib/util';
