+++ src/mahoji/lib/abstracted_commands/cancelGEListingCommand.ts
@@ -1,9 +1,44 @@
+import { UserError } from '@oldschoolgg/toolkit/dist/lib/UserError';
 import { Bank } from 'oldschooljs';
 
 import { GrandExchange } from '../../../lib/grandExchange';
 import { prisma } from '../../../lib/settings/prisma';
 import { makeTransactFromTableBankQueries } from '../../../lib/tableBank';
 import { logError } from '../../../lib/util/logError';
+
+export async function cancelUsersListings(user: MUser) {
+	const activeListings = await prisma.gEListing.findMany({
+		where: {
+			user_id: user.id,
+			quantity_remaining: {
+				gt: 0
+			},
+			fulfilled_at: null,
+			cancelled_at: null
+		},
+		include: {
+			buyTransactions: true,
+			sellTransactions: true
+		},
+		orderBy: {
+			created_at: 'desc'
+		}
+	});
+	// Return early if no active listings.
+	if (activeListings.length === 0) {
+		return true;
+	}
+	// Let's terminate all listings:
+	for (const listing of activeListings) {
+		const result = await cancelGEListingCommand(user, listing.userfacing_id);
+		if (!result.startsWith('Successfully cancelled your listing')) {
+			const err = `Failed to cancel ${user.usernameOrMention}'s listings: ${result}`;
+			logError(new Error(err));
+			throw new UserError(err);
+		}
+	}
+	return true;
+}
 
 export async function cancelGEListingCommand(user: MUser, idToCancel: string) {
 	if (GrandExchange.locked) {
